myproject/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-formtools==2.5.1
django-oauth-toolkit==3.0.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.9
jwcrypto==1.5.6
msgpack==1.1.0
oauthlib==3.2.2
phonenumbers==8.13.45
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
